Accept id set to None
Existing code was expecting the "id" field to be a non "None" value
whenever it was present, so when the "None" value was set we would not
generate a new UUID and the creation of the OVO would fail because the
"id" field could not be "None".

This patch adds support for UUID generation when "id" is set to "None".