Fix read the docs building error
The build in readthedocs is failing with error:

  error in setup command: Error parsing /tmp/pip-xIRoVN-build/setup.cfg:
  TypeError: super() argument 1 must be type, not None

Since this seems to be a problem when installing the full Cinder package
this patch uses a faker for the Cinder package to satisfy the
requirements of the cinderlib package and in the Sphinx configuration
also fakes all other libraries that get pulled with the Cinder package
and that cinderlib expects to have.

Ensure we are using the same Sphinx version (1.6.5) everywhere.