Add support for default project_id and user_id
The user_id and project_id fields for our volumes and snapshots had a
very poor default value, they were defaulting to None.  And the only way
we had to change them was to provide an override for these fields when
creating the resource.

This patch adds a better default for these values ("cinderlib") and also
supports setting a different default on the setup process passing
project_id and user_id parameters to the "setup" method.