Improve stats documentation
In the documentation we mention that collecting the stats is an
expensive operation and that we can read from the cache, but we don't
mention that reading from the cache is the default or how we can force a
refresh.

This patch modifies the backends topic in the documentation to include
this information.