Add metadata plugin mechanism
Until now we only supported storing resources in memory, and that
mechanism was tightly coupled with everything else.

This patch decouples the metadata storing into a plugin mechanism using
'Python entrypoints.  The default storing mechanism is still in memory,
and for now it's the only mechanism available.

The JSON serialization mechanism has not yet been adapted to the
metadata plugin mechanism, so it will only work with the default plugin.