Fix name & description inconsistency
There was an inconsistency on the name and description fields, since you
could pass them to the create volume and snapshot methods but then you
couldn't use them to see the values of the objects and you had to use
display_name and display_description instead.

This patch fixes this and renames the old OVO name property to
name_in_storage.