Add DB persistence plugin
This patch adds 2 new metadata storage options based on Cinder's
SQLAlchemy support:

- DB (db entrypoint)
- SQLite in memory DB (db_memory entrypoint)

Using the DB metadata storage we can use any of the DBMSs supported by
the Cinder project.

The benefit of using the "db_memory" plugin is that some drivers that
may fail with the default "memory" plugin may work with it if the driver
is accessing the DB directly from the driver and Cinderlib hasn't mokey
patched those DB calls.

The JSON serialization mechanism that got broken when we introduced the
metadata plugin mechanism has not been fixed in this patch.