Automatically set availability_zone in Volume
In the Volume cinderlib object we store the backend in the
availability_zone field and we were passing this field on the
Backend.create_volume method to the objects.Volume.

This patch changes this so the create_volume doesn't have to care about
it and it's the Volume.__init__ method that takes care of setting the
field for all sources (kwargs, volume, ovo).