Add dump and dumps methods
Now that we don have everything in memory we want to differentiate
between serializing what we have in memory and ensuring we serialize
even the lazy loading attributes.

So we have now methods to serialize whatever we have in memory -with the
json method-, and making sure the lazy loaded attributes have been
loaded -with dump method-.

Lazy loading is not recursive, it will only load fields from the object
where "dump" has been called.

If a lazy loadable attribute is already loaded it is not refreshed.