Add minimum serialization method to objects
We have "json" and "dump" serialization mechanisms, but in both cases
serialized data presents internal information used for backward
compatibility of the objects.  This is an inconvenience when we just
want the fields from the objects.

This patch adds a new serialization method called "to_dict" that is
meant to return only the data fields relevant to the user.