Improve lazy loaded volumes list
Currently lazy loaded `volumes` property list will get out of sync with
the contents of the metadata store as soon as we create a volume.  This
happens even when there's only one process running.

This patch fixes this by adding created volumes to the internal
`_volumes` list if it has been loaded from the metadata store.