Support Ceph/RBD in functional tests
Cinder's RBD driver cannot delete a snapshot if it has a dependent
volume.  That is, a volume that was created from the snapshot.  This is
making the functional test run fail on cleanup.

To solve this problem we now clean up volumes in reverse order of
creation.

To test the RBD driver we need to install the `ceph-common` package,
which install system required CLI tools as well as Python libraries.  To
avoid installing these libraries at the system level and then in the
testing virtual environment we allow using Site Packages in the Tox
virtual environment.