Fix multipath
We were no longer able to do multipathing after changes to the way we
store the connector and connection information in he Connector class.

This patch fixes this by taking into account the new location of the
connector_dict information to detect when we have to do a multipath
attachment.