Fix github README file
Github doesn't support the "include" directive, so when we fixed the
readme for the PyPi package we broke the github readme.  Now it's only
presenting the demo.

To fix this we revert to having the whole readme in the README.rst file
and in setup.py we hardcode the removal of the demo section.