tests: Support older versions of lsblk
Our functional tests use lsblk to get the size of a device, and we were
using the JSON output that is available on v2.28.2.

With this patch we change how we use lsblk and stop requesting JSON
output and read normal output.  Making us able to use older version of
lsblk.

Tested with lsblk version 2.23.2