Support loading object without configured backend
Global configuration option "fail_on_missing_backend" allows us to load
a volume from the metadata persistence storage even if the backend is
not configured.

This is usefull if we share the metadata storage and we want to load a
volume that is already connected to do just the attachment.