Fix README on Github
Now that we have multiple README files, one for Docker Hub, and another
for Github, we have the wrong file rendered by Github because it chooses
the Markdown instead of the RST file (alphabetical order [1]).

We rename the README.md to README to resolve this, as this seems to help
us get the right file in both cases.

[1] https://github.com/github/markup/issues/360#issuecomment-55700463