Workaround for Python 2 getaddrinfo bug
Python 2 has a bug on the builtin getaddrinfo method that incorrectly
fails when receiving a unicode string as the `port` parameter.  It only
accepts `str` and `int` values.

With the workaround we wrap the builtin and convert the port to a `str`
when it's a `unicode`.

One of the drivers affected by this bug is the QNAP driver.