Fix README example
After adding the metadata persistence plugin system the example in the
README doesn't work anymore.

The reason is that the serialization doesn't save the local attach
information, so when we deserialize it we don't know the volume has a
local attachment.

Update the example to show how to do the detach.