Stop setting attributes on CONF
We are currently setting attributes directly on CONF, which works if you
just access those attributes because the default Oslo Config __getattr__
method doesn't get called, but that value will not be used in any other
case.

Such cases is when you do something like `dict(CONF.database)`.

To prevent this from happening we now use method `set_override` to set
configuration option.  This has the added benefit of checking that the
configuration option exists and to run validation on the provided value.