Support requests and requests-kerberos
Depending on the version of requests we are using we could be failing on
setup with error:

  AttributeError: 'module' object has no attribute 'urllib3'

This happens when we are supressing ssl_warnings, as we access the the
urllib3 module through the requests, and depending on the library being
used could not be present.

This patch ensures we call the disabling of warnings directly on
urllib3.