docs: Update LVM configuration
Our documentation uses iscsi_protocol and iscsi_helper to initialize the
LVM backend in multiple places, but this configuration option was being
deprecated in Pike and was removed in Queens.

This patch replaces these configuration options with their target_*
counterparts so the documentation is valid for all cinderlib supported
releases.