Support pool-aware drivers
This patch adds pool support to cinderlib.

The Backend class now has the property `pool_names` with available pools
in the driver.

When creating a volume we can specify `pool_name` to use a specific
pool.

Pool is stored as part of the `host` field in volumes.  The `host` field
now follows Cinder's naming "host@backend#pool".