Support NBD RBD attachments
RBD connections were done using the kernel RBD module, which is always
behind, feature-wise, to the Ceph clusters.

Being behind means that you have to disable some features on the client
by editing the ceph.conf file, which is quite inconvenient.

This patch uses of the NBD driver when present in the system, and since
it uses librbd we don't have those feature issues.

If NBD is not present we use the kernel RBD module instead.