Add support for MultiOpt parameters
Some drivers use MultiOpt parameters, which are parameters that can be
present multiple times in a configuration file and which the parser
joins together into a single list.

This is not a supported feature of normal .INI file parsers, as the last
configuration option overwrites the old value.

This patch adds a "hack" to make it work, so we now support MultiOpt
parameters that must be apssed to cinderlib as a list or a tuple.