Workaround for driver stats bug
Some drivers don't have a working stats cache as they should, and so
they return an empty dictionary when we try to retrieve their cached
stats by passing refresh=False.

On such drivers is HPE's 3PAR.

This patch works around this by checking returned value and call it
again with refresh=True if the backend doesn't work as it should.