Fix RBD connector
RBD nos-brick connector has a bug where it only detaches the first
volume if we are running cinderlib in a container as root and the host
doesn't have the ceph-common package installed.

The library says it has detached it, but in reality it hasn't, so we end
up with a /dev/rbdX device leftover in our system.

This patch fixes this by taking into account that directories and
symlinks may already exist.

Closes-Bug: #1819705
Change-Id: Ib61acdd4c469ac4316e2f5e518301c44dd3a6321