Enable publish-to-pypi Zuul job
This patch enables the publish-to-pypi Zull job and for that removes the
git repository from the test requirements and adds it instead to the Tox
file.

It needs to split the README file into 2 parts, one that is used for
PyPi, and the other for renderers that support the raw directive.

Change-Id: Id029cedef513753700a10bcae761fdc1807964e1