Support setting attach_mode
We cannot set the attach_mode when connecting or attaching a volume
because the "Connection.connect" method is blindly using
"attach_mode='rw'".

This patch makes it possible to pass the "attach_mode" to both the
volume attach and the volume connect operation.  Setting the
"access_mode" in the connection info, like Cinder does.

Closes-Bug: #1819706
Change-Id: Ie1e82d52ca5f8f1be1249756963f72d545a26288