Fix error_deleting status in vol and snap
When an error happens deleting a volume or a snapshot the status change
to "error_deleting" is not persisted in the metadata storage.

The reason for this is that we were creating a new "status" attribute in
cinderlib's objects instead of setting it in the existing OVO's
attribute.

This patch fixes this for both volumes and snapshots.

Closes-Bug: #1819708
Change-Id: I69f703c1d46dc4b19120e786c43998d8e5aefb21